lint
SCCS-vsn: sys/vax/uba/ts.c 4.33
SCCS-vsn: sys/net/if.c 4.21
SCCS-vsn: sys/net/raw_usrreq.c 4.20
SCCS-vsn: sys/net/route.c 4.13
SCCS-vsn: sys/netinet/in.c 4.6
SCCS-vsn: sys/netinet/ip_icmp.c 4.21
SCCS-vsn: sys/netinet/ip_input.c 1.53
SCCS-vsn: sys/netinet/ip_output.c 1.39
SCCS-vsn: sys/netinet/tcp_debug.c 4.7
SCCS-vsn: sys/netinet/tcp_input.c 1.79
SCCS-vsn: sys/netinet/tcp_subr.c 4.32
SCCS-vsn: sys/netinet/tcp_timer.c 4.27
SCCS-vsn: sys/netinet/tcp_usrreq.c 1.67
SCCS-vsn: sys/netinet/udp_usrreq.c 4.36
SCCS-vsn: sys/vax/uba/uda.c 4.13
SCCS-vsn: sys/kern/kern_proc.c 4.42
SCCS-vsn: sys/kern/kern_sig.c 5.9
SCCS-vsn: sys/kern/vfs_lookup.c 4.27
SCCS-vsn: sys/ufs/ffs/ufs_lookup.c 4.27
SCCS-vsn: sys/ufs/ufs/ufs_lookup.c 4.27
SCCS-vsn: sys/vax/vax/Locore.c 4.21
SCCS-vsn: sys/vax/vax/mem.c 4.7
SCCS-vsn: sys/vax/uba/dn.c 4.11
SCCS-vsn: sys/vax/uba/tm.c 4.58
SCCS-vsn: sys/vax/uba/ts.c 4.34
SCCS-vsn: sys/vax/uba/ut.c 4.21
SCCS-vsn: sys/vax/mba/ht.c 4.32
SCCS-vsn: sys/vax/mba/mt.c 4.13

+++ usr/src/sys/netinet/udp_usrreq.c
@@ -1,4 +1,4 @@
-/*	udp_usrreq.c	4.35	82/10/17	*/
+/*	udp_usrreq.c	4.36	82/10/17	*/
 
 #include "../h/param.h"
 #include "../h/dir.h"
@@ -151,7 +151,7 @@ udp_ctlinput(cmd, arg)
 
 	default:
 		sin = &((struct icmp *)arg)->icmp_ip.ip_dst;
-		in_pcbnotify(&udb, sin, inetctlerrmap[cmd], udp_abort);
+		in_pcbnotify(&udb, sin, (int)inetctlerrmap[cmd], udp_abort);
 	}
 }
